much ado about nothing
the current setup for reading a line would hang for user input if the line was empty rather than simply printing the prompt again.

so, naturally to fix this minor bug, i completely reworked an entire code file, and integrated the feature into libClame, and updated the source tree to use the latest version of libClame.

i suffer so you don't have to.