added support for direct assembly compilation
however there is a big bug that causes the compiled binary to try to access code below the start of the array. i'm trying to compensate for that partially by adding extra space below the array but this smells like a bug in the optimiser.

otherwise, there is one more example program and some reformatting of dialogues and stuff.